% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone_esri.R
\name{esri_to_sgdf}
\alias{esri_to_sgdf}
\title{Convert a base64-encoded gzipped Esri ASCII grid to an object of class SpatialGridDataFrame}
\usage{
esri_to_sgdf(isochrone_esri)
}
\arguments{
\item{isochrone_esri}{a response object from a \link[walkalytics]{isochrone_esri} call to the \code{walkalytics} isochrone API.}
}
\description{
\code{esri_to_sgdf} converts a response object from a \link[walkalytics]{isochrone_esri} call to the \code{walkalytics} isochrone API to
   an object of class \link[sp]{SpatialPixelsDataFrame-class}.
}
\examples{
\dontrun{
isochrone_esri(x = 895815, y = 6004839, key = "abcd1234") \%>\% esri_to_sgdf()
}

}
